Migrating command and script plugins to a true plugin
