+++ core/src/main/java/com/dtolabs/rundeck/core/execution/BaseCommandExec.java
@@ -1,13 +1,13 @@
 package com.dtolabs.rundeck.core.execution;
 
 public interface BaseCommandExec {
-    String getArgString();
-    String getAdhocRemoteString();
-    String getAdhocLocalString();
-    String getAdhocFilepath();
-    Boolean getAdhocExecution();
-    String getScriptInterpreter();
-    String getFileExtension();
-    Boolean getInterpreterArgsQuoted();
-    Boolean getExpandTokenInScriptFile();
+    default String getArgString() {return null;};
+    default String getAdhocRemoteString(){return null;};
+    default String getAdhocLocalString(){return null;};
+    default String getAdhocFilepath(){return null;};
+    default Boolean getAdhocExecution(){return null;};
+    default String getScriptInterpreter(){return null;};
+    default String getFileExtension(){return null;};
+    default Boolean getInterpreterArgsQuoted(){return null;};
+    default Boolean getExpandTokenInScriptFile(){return null;};
 }
