+++ core/src/main/java/com/dtolabs/rundeck/core/execution/ExecCommand.java
@@ -1,0 +1,5 @@
+package com.dtolabs.rundeck.core.execution;
+
+public interface ExecCommand extends BaseCommandExec {
+    static final String EXEC_COMMAND_TYPE = "exec-command";
+}
