+++ core/src/main/java/com/dtolabs/rundeck/core/execution/ScriptCommand.java
@@ -1,0 +1,5 @@
+package com.dtolabs.rundeck.core.execution;
+
+public interface ScriptCommand extends BaseCommandExec{
+    static final String SCRIPT_COMMAND_TYPE = "script-command";
+}
