+++ core/src/main/java/com/dtolabs/rundeck/core/execution/ScriptFileCommand.java
@@ -1,0 +1,5 @@
+package com.dtolabs.rundeck.core.execution;
+
+public interface ScriptFileCommand extends BaseCommandExec{
+    static final String SCRIPT_FILE_COMMAND_TYPE = "script-file-command";
+}
