+++ core/src/main/java/com/dtolabs/rundeck/core/execution/workflow/steps/node/NodeStepExecutionService.java
@@ -68,9 +68,6 @@ public class NodeStepExecutionService
     static {
         Map<String, Class<? extends NodeStepExecutor>> map = new HashMap<>();
         map.put(ExecNodeStepExecutor.SERVICE_IMPLEMENTATION_NAME, ExecNodeStepExecutor.class);
-        map.put(ScriptNodeStepExecutor.SERVICE_IMPLEMENTATION_NAME, ScriptNodeStepExecutor.class);
-//        map.put(ScriptFileNodeStepExecutor.SERVICE_IMPLEMENTATION_NAME, ScriptFileNodeStepExecutor.class);
-//        map.put(ScriptURLNodeStepExecutor.SERVICE_IMPLEMENTATION_NAME, ScriptURLNodeStepExecutor.class);
         PRESET_PROVIDERS = Collections.unmodifiableMap(map);
     }
